crawling/requirements.txt
PyMuPDF==1.24.10
Pillow==10.4.0

